requirements.txt
flask==3.0.0
python-docx==1.1.0
openai>=1.57.0
pydantic>=2.0.0
python-dotenv==1.0.0
werkzeug==3.0.1
PyPDF2==3.0.1
python-pptx==0.6.23
